overhaul response tab: - prettier error message - improved styling - show only views with actual data
